+++ src/chrome/map.html
@@ -7,6 +7,6 @@
   </head>
   <body style="background: lightgrey">
     <div id="container" style="width: 90vw; height: 90vh; background: white"></div>
-    <script src="../../dist/graph.bundle.js"></script>
+    <script src="../../dist/main.bundle.js"></script>
   </body>
 </html>
