+++ webpack.config.js
@@ -1,11 +1,11 @@
 const path = require('path');
 
 module.exports = {
-  entry: './src/sigma/graph.js',
+  entry: './main.js',
   output: {
-    filename: 'graph.bundle.js',
+    filename: 'main.bundle.js',
     path: path.resolve(__dirname, 'dist'),
   },
   mode: 'development',
   devtool: 'source-map'
-};``
+};
